mcp_servers/code/pyproject.toml
[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["mcp_servers/code_execution_server/tests"]
pythonpath = ["mcp_servers/code_execution_server"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["mcp_servers/code_execution_server/tools"]
omit = ["*/tests/*", "*/test_*.py", "*/__pycache__/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[project]
name = "mercor-rls-code-execution"
version = "0.1.0"
description = "Mercor RLS Code Execution MCP"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "pip>=24.0",  # Ensure pip is available for subprocess calls
    "aioboto3>=15.4.0",
    "aiobotocore>=2.25.0",
    "asyncer>=0.0.9",
    "asyncpg>=0.30.0",
    "datadog-api-client>=2.44.0",
    "fastmcp>=2.12.4",
    "httpx>=0.27.0",
    "litellm>=1.77.7",
    "loguru>=0.7.3",
    "mcp-schema",
    "pydantic-settings>=2.11.0",
    "redis>=6.4.0",
    "matplotlib>=3.8.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scipy>=1.11.0",
    "seaborn>=0.13.0",
    "statsmodels>=0.14.0",
    "plotly>=5.15.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "duckdb>=1.0.0",
    # Web scraping
    "beautifulsoup4>=4.12.0",  # HTML parsing (import as bs4)
    "html5lib>=1.1",           # HTML5 parser for BeautifulSoup
    # Document generation
    "reportlab>=4.0.0",    # PDF (feature-rich)
    "fpdf2>=2.7.0",        # PDF (lightweight, agents often try this first)
    "pypdf>=5.0.0",        # PDF (reading/manipulation)
    "pdfplumber>=0.10.0",  # PDF extraction with table support
    "openpyxl>=3.1.0",     # Spreadsheets
    "python-pptx>=0.6.21", # Presentations
    "python-docx>=1.1.0",  # Documents
]

[dependency-groups]
dev = [
    "basedpyright>=1.31.6",
    "ruff>=0.14.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]
medicine = ["pydicom>=2.4.0"]
finance = [
    "numpy-financial>=1.0.0",   # IRR, NPV, PMT financial calculations
    "yfinance>=0.2.0",          # Yahoo Finance market data
    "pandas-datareader>=0.10.0",  # Remote data access for pandas
]
data-science = [
    "pymupdf>=1.24.0",  # Fast PDF library (import as fitz)
    "PyPDF2>=3.0.0",    # Legacy PDF library agents sometimes use
    "mpmath>=1.3.0",    # Arbitrary precision math
]
scicomp = [
    # Fully Native Python scientific computing libraries
    "biopython>=1.84",           # Biological computation (covers 3+ tools)
    "requests>=2.31.0",          # HTTP library (covers 5+ APIs)
    "openmm>=8.0.0",             # Molecular dynamics simulation
    "pyhmmer>=0.10.0",           # Python bindings for HMMER (replaces binary phmmer)
    "particle>=0.24.0",          # Particle physics data
]

[tool.basedpyright]
typeCheckingMode = "recommended"
reportUnusedCallResult = "none"
failOnWarnings = false
exclude = [
    "packages",
    "modal_labs.py",
    "modal_helpers.py",
    ".venv",
    ".git",
    ".tmp",
    "venv",
    "mcp_servers",
    "smoke_test",
]

[tool.ruff.lint]
# Enable linting rules
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "W",  # pycodestyle warnings
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]

ignore = [
    "E501", # Line too long
]

[tool.uv.sources]
fastmcp = { git = "https://github.com/jlowin/fastmcp.git", rev = "bc2f601" }
mcp-schema = { path = "packages/mcp_schema" }
